requirements.txt
fastapi
uvicorn[standard]
pytesseract
Pillow
python-multipart
jinja2
fastapi-proxiedheadersmiddleware
# --extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
diffusers